Add `verbose` flag to CLI and Python API (#32)
* update pre-commit hooks

* ruff ignore PLW2901

* change lax_steps and lax_runs exception type from AssertionError to ValueError

* add CLI flag --verbose and verbose: bool = False kwarg to read/write and reduce functions

* parametrize verbose kwarg in test_load_tb_events_strict(), test_reduce_events(), test_write_tb_events(), test_write_data_file()

* pyproject.toml fix tool.setuptools.packages.find include/exclude

* remove needless dict.keys()

* add tqdm to package deps

* pandas will drop .xls support since xlwt/xlrd are unmaintained so we do too

df.to_excel('out.xls') emits FutureWarning: As the xlwt package is no longer maintained, the xlwt engine will be removed in a future version of pandas. This is the only engine in pandas that supports writing in the xls format. Install openpyxl and write to an xlsx file instead. You can set the option io.excel.xls.writer to 'xlwt' to silence this warning. While this option is deprecated and will also raise a warning, it can be globally set and the warning suppressed.

* move loaded-data report from main() into load_tb_events()

* set verbose=True in example notebooks